requirements.txt
python-dotenv
pandas
numpy
matplotlib
seaborn
pymongo
pymongo[srv]==3.6
certifi
#-e .